+++ asylum-assistant/src/app/app.module.ts
@@ -4,7 +4,6 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { CountriesSelectComponent } from './home/countries-select/countries-select.component';
 import { FormsModule } from '@angular/forms';
 import { FormComponent } from './form/form.component';
 import { IntroComponent } from './intro/intro.component';
@@ -15,7 +14,6 @@ import { HomeComponent } from './home/home.component';
 @NgModule({
   declarations: [
     AppComponent,
-    CountriesSelectComponent,
     FormComponent,
     IntroComponent,
     OptionsComponent,
