--- asylum-assistant/src/app/home/countries-select/countries-select.component.css
@@ -1,17 +1,0 @@
-img.countryImg {
-  width: 48px;
-}
-
-/*.list-group{
-  max-height: 80%;
-  margin-bottom: 10px;
-  overflow:scroll;
-  -webkit-overflow-scrolling: touch;
-}*/
-
-.list-group {
-  /*max-width: 25%;*/
-  margin: auto;
-  overflow:scroll;
-  -webkit-overflow-scrolling: touch;
-}
