--- asylum-assistant/src/app/home/countries-select/countries-select.component.html
@@ -1,15 +1,0 @@
-<div class="container text-center">
-  <h2 for="welcome">Welcome to Spain</h2>
-  <p>Choose Your Country of Origin</p>
-
-  <div>
-    <ul class="list-group" [style.height.px]="innerHeight">
-      <li *ngFor="let country of countriesList" class="list-group-item">
-        <a [routerLink]="['intro']" to [routerLink]="['/intro']">
-          <img class="countryImg" [src]="'/assets/img/countries/' + country.flag + '.png'" /> {{ country.name }}
-        </a>
-      </li>
-    </ul>
-  </div>
-</div>
-<router-outlet></router-outlet>
