+++ asylum-assistant/src/app/home/home.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, HostListener, OnInit } from '@angular/core';
+import {default as countries} from '../../assets/json/countries.json';
 
 @Component({
   selector: 'app-home',
@@ -6,10 +7,25 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
+  selectedCountry: number = -1;
+
+  countriesList: Array<any> = countries;
+
+  public innerHeight: any;
 
   constructor() { }
 
   ngOnInit(): void {
+    this.selectedCountry = -1;
+    this.innerHeight = window.innerHeight * .75;
   }
 
+  public changeValue() {
+    this.selectedCountry = this.countriesList.filter(a => a.defaultSelected)[0].value;
+  }
+
+  @HostListener('window:resize', ['$event'])
+  onResize(event) {
+    this.innerHeight = window.innerHeight * .75;
+  }
 }
