+++ asylum-assistant/src/app/intro/intro.component.html
@@ -3,5 +3,6 @@
     <source src="assets/vid.mp4" type="video/mp4">
   </video>-->
 
-  <a href="#" class="btn btn-success text-right">Next</a>
+  <a [routerLink]="['options']" to [routerLink]="['/options']" class="btn btn-success text-right">Next</a>
 </div>
+<router-outlet></router-outlet>
